set up foundations for mobile pages. Reorganization of file locations
