+++ src/components/landing/desktop/about.tsx
