+++ src/components/landing/desktop/footer.tsx
